requirements.txt
fastapi[standard]==0.116.1
yfinance==0.2.65
python-dotenv==1.1.0
pymongo==4.12.1
uvicorn
setuptools
passlib==1.7.4
python-jose==3.4.0
bcrypt==4.3.0
requests==2.32.3
redis==4.3.4